Fixing User Email Verification and gitignore
The user verification was not forcing users to verify through email because of the vendor file modifications not being pushed to master. The gitignore has been fixed so that the file changes to only the two relevant php files will be pushed to the master. The remaining vendor files will be ignored as they should be, which should prevent problems with those files and merge conflicts in the future, such as the one that Ramses and I encountered today.